chore: decompose geometry/entities.py, centralize diagnostic utils, and extract minimizer evaluators
